+++ app/Http/Controllers/Api/Resource/Photos/IndexController.php
@@ -52,7 +52,8 @@ class IndexController extends Controller
      */
     public function __invoke(IndexRequest $request, $resource)
     {
-        $photos = $resource->photos()->latest()->with('user', 'comment')->paginate(20);
+        $photos = $resource->photos()->latest()->paginate(20);
+        $photos->load(['user', 'comment']);
         return response()->json($photos,200);
     }
 }
