+++ tests/Feature/Stories/UserCanViewOwnedPhotosTest.php
@@ -6,12 +6,14 @@ namespace Tests\Feature\Stories;
 
 use App\Photo;
 use App\User;
+use Illuminate\Foundation\Testing\DatabaseMigrations;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
 
 class UserCanViewOwnedPhotosTest extends TestCase
 {
-    use RefreshDatabase;
+    use DatabaseMigrations;
+
     /**
      * @test testIndexUsersPhotosApi
      */
